pia/requirements.txt
av==13.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
edge_impulse_linux==1.0.10
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
picamera2==0.3.21
pidng==4.0.9
piexif==1.1.3
pillow==11.0.0
psutil==6.1.0
PyAudio==0.2.14
python-prctl==1.8.1
requests==2.32.3
RPi.GPIO==0.7.1
simplejpeg==1.7.6
six==1.16.0
urllib3==2.2.3
v4l2-python3==0.3.5
